アクションのコメント整備

+++ exam-com/src/main/java/com/example/action/model/MCodeDeleteAction.java
@@ -11,23 +11,33 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * コードマスタ削除
+ *
+ * @author emarfkrow
+ */
 public class MCodeDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** コードマスタ削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
-        Map<String, Object> map = new HashMap<String, Object>();
+        // 主キーが不足していたらエラー
+        Object codeNm = postJson.get("codeNm");
+        if (codeNm == null) {
+            codeNm = postJson.get("MCode.codeNm");
+        }
+        if (codeNm == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         MCode e = FormValidator.toBean(MCode.class.getName(), postJson);
-
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
-        map.put("INFO", Messages.get("info.delete"));
 
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("INFO", Messages.get("info.delete"));
         return map;
     }
