+++ exam-com/src/main/java/com/example/action/model/MShokuiDeleteAction.java
@@ -11,23 +11,33 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * 職位マスタ削除
+ *
+ * @author emarfkrow
+ */
 public class MShokuiDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** 職位マスタ削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
-        Map<String, Object> map = new HashMap<String, Object>();
+        // 主キーが不足していたらエラー
+        Object shokuiId = postJson.get("shokuiId");
+        if (shokuiId == null) {
+            shokuiId = postJson.get("MShokui.shokuiId");
+        }
+        if (shokuiId == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         MShokui e = FormValidator.toBean(MShokui.class.getName(), postJson);
-
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
-        map.put("INFO", Messages.get("info.delete"));
 
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("INFO", Messages.get("info.delete"));
         return map;
     }
