+++ exam-com/src/main/java/com/example/action/model/MUserDeleteAction.java
@@ -11,23 +11,33 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * ユーザマスタ削除
+ *
+ * @author emarfkrow
+ */
 public class MUserDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** ユーザマスタ削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
-        Map<String, Object> map = new HashMap<String, Object>();
+        // 主キーが不足していたらエラー
+        Object userId = postJson.get("userId");
+        if (userId == null) {
+            userId = postJson.get("MUser.userId");
+        }
+        if (userId == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         MUser e = FormValidator.toBean(MUser.class.getName(), postJson);
-
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
-        map.put("INFO", Messages.get("info.delete"));
 
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("INFO", Messages.get("info.delete"));
         return map;
     }
