+++ exam-com/src/main/java/com/example/action/model/TEntityHisDeleteAction.java
@@ -11,23 +11,54 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * エンティティ履歴削除
+ *
+ * @author emarfkrow
+ */
 public class TEntityHisDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** エンティティ履歴削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
-        Map<String, Object> map = new HashMap<String, Object>();
+        // 主キーが不足していたらエラー
+        Object sosenId = postJson.get("sosenId");
+        if (sosenId == null) {
+            sosenId = postJson.get("TEntityHis.sosenId");
+        }
+        if (sosenId == null) {
+            throw new OptLockError("error.cant.delete");
+        }
+        Object oyaSn = postJson.get("oyaSn");
+        if (oyaSn == null) {
+            oyaSn = postJson.get("TEntityHis.oyaSn");
+        }
+        if (oyaSn == null) {
+            throw new OptLockError("error.cant.delete");
+        }
+        Object entitySn = postJson.get("entitySn");
+        if (entitySn == null) {
+            entitySn = postJson.get("TEntityHis.entitySn");
+        }
+        if (entitySn == null) {
+            throw new OptLockError("error.cant.delete");
+        }
+        Object historySn = postJson.get("historySn");
+        if (historySn == null) {
+            historySn = postJson.get("TEntityHis.historySn");
+        }
+        if (historySn == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         TEntityHis e = FormValidator.toBean(TEntityHis.class.getName(), postJson);
-
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
-        map.put("INFO", Messages.get("info.delete"));
 
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("INFO", Messages.get("info.delete"));
         return map;
     }
