+++ exam-com/src/main/java/com/example/action/model/TOyaDeleteAction.java
@@ -11,23 +11,40 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * 親削除
+ *
+ * @author emarfkrow
+ */
 public class TOyaDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** 親削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
-        Map<String, Object> map = new HashMap<String, Object>();
+        // 主キーが不足していたらエラー
+        Object sosenId = postJson.get("sosenId");
+        if (sosenId == null) {
+            sosenId = postJson.get("TOya.sosenId");
+        }
+        if (sosenId == null) {
+            throw new OptLockError("error.cant.delete");
+        }
+        Object oyaSn = postJson.get("oyaSn");
+        if (oyaSn == null) {
+            oyaSn = postJson.get("TOya.oyaSn");
+        }
+        if (oyaSn == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         TOya e = FormValidator.toBean(TOya.class.getName(), postJson);
-
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
-        map.put("INFO", Messages.get("info.delete"));
 
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("INFO", Messages.get("info.delete"));
         return map;
     }
