+++ exam-com/src/main/java/com/example/action/model/TSosenDeleteAction.java
@@ -11,23 +11,33 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * 祖先削除
+ *
+ * @author emarfkrow
+ */
 public class TSosenDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** 祖先削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
-        Map<String, Object> map = new HashMap<String, Object>();
+        // 主キーが不足していたらエラー
+        Object sosenId = postJson.get("sosenId");
+        if (sosenId == null) {
+            sosenId = postJson.get("TSosen.sosenId");
+        }
+        if (sosenId == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         TSosen e = FormValidator.toBean(TSosen.class.getName(), postJson);
-
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
-        map.put("INFO", Messages.get("info.delete"));
 
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("INFO", Messages.get("info.delete"));
         return map;
     }
